+++ src/ra/building.h
@@ -343,7 +343,7 @@ class BuildingClass : public TechnoClass {
   void Repair(int control) override;
   void Sell_Back(int control) override;
   RadioMessageType Receive_Message(RadioClass* from, RadioMessageType message,
-                                   long& param) override;
+                                   int32_t& param) override;
   void AI() override;
   void Assign_Target(TARGET target) override;
   virtual bool Toggle_Primary();
