+++ src/ra/cell.cc
@@ -239,7 +239,7 @@ TechnoClass* CellClass::Cell_Techno(int x, int y) const {
   ObjectClass* object;
   COORDINATE click;  // Coordinate of click relative to cell corner.
   TechnoClass* close = nullptr;
-  long distance = 0;  // Recorded closest distance.
+  int32_t distance = 0;  // Recorded closest distance.
 
   /*
   **	Create a coordinate value that represent the pixel location within the
@@ -253,7 +253,7 @@ TechnoClass* CellClass::Cell_Techno(int x, int y) const {
     while (object) {
       if (object->Is_Techno()) {
         COORDINATE coord = Coord_Fraction(object->Center_Coord());
-        long dist = Distance(coord, click);
+        int32_t dist = Distance(coord, click);
         if (!close || dist < distance) {
           close = dynamic_cast<TechnoClass*>(object);
           distance = dist;
@@ -1850,7 +1850,7 @@ void CellClass::Adjust_Threat(HousesType house, int threat_value) {
  * HISTORY: * 05/16/1995 JLB : Created. * 02/20/1996 JLB : Takes into account
  *the ore type.                                         *
  *=============================================================================================*/
-long CellClass::Tiberium_Adjust(bool pregame) {
+int32_t CellClass::Tiberium_Adjust(bool pregame) {
   assert(static_cast<unsigned>(Cell_Number()) <= MAP_CELL_TOTAL);
   if ((Overlay != OVERLAY_NONE) &&
       (OverlayTypeClass::As_Reference(Overlay).Land == LAND_TIBERIUM)) {
@@ -1909,7 +1909,7 @@ long CellClass::Tiberium_Adjust(bool pregame) {
     } else {
       OverlayData = static_cast<unsigned char>(_adj[count]);
     }
-    return (static_cast<long>(OverlayData + 1)) * value;
+    return (static_cast<int32_t>(OverlayData + 1)) * value;
   }
 
   return 0;
